Fixed bug UtilityItem will now be removed from inventory when completed used.

+++ Assets/Scripts/Items/Utility Items/UtilityItem.cs
@@ -30,7 +30,7 @@ namespace Items.UtilityItems
         {
             get
             {
-                return initialValue * (float) Uses / Hardness;
+                return initialValue * Uses / Hardness;
             }
         }
